backend/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
pydantic==2.6.4
SQLAlchemy==2.0.29
PyGithub==2.3.0
google-generativeai==0.4.1
python-dotenv==1.0.1
itsdangerous==2.1.2
httpx==0.27.0
bandit==1.7.8
starlette==0.37.2
